+++ swagger_parser/lib/src/parser/model/open_api_info.dart
