+++ swagger_parser/lib/src/parser/swagger_parser_core.dart
@@ -1,0 +1,7 @@
+///
+library swagger_parser_core;
+
+export 'config/parser_config.dart';
+export 'exception/parser_exception.dart';
+export 'model/open_api_info.dart';
+export 'parser/open_api_parser.dart';
