+++ swagger_parser/lib/src/utils/utils.dart
@@ -1,10 +1,10 @@
 import 'dart:io';
 
 import '../generator/models/generation_statistics.dart';
-import '../generator/models/open_api_info.dart';
 import '../generator/models/programming_language.dart';
 import '../generator/models/universal_data_class.dart';
 import '../generator/models/universal_type.dart';
+import '../parser/model/open_api_info.dart';
 import '../utils/case_utils.dart';
 
 const _green = '\x1B[32m';
